relocated authenticate methods and combined with signed params logic
