+++ config/environment.rb
@@ -6,7 +6,7 @@ require 'token_maker'
 require 'locale_selector'
 require 'markdown_wrapper'
 require 'user_session_management'
-require 'use_signed_params'
+require 'authenticate_methods'
 require 'contracts_not_required'
 require 'require_recent_signin'
 require 'json_serialize'
